transform-service/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-bigquery==3.*